Update excelTranPer.py
Message boxes and choose file option for simplicity but still needs a way to write to a file so I must replace all of those prints with a write to. I need to figure out how to make all of that a class and maybe have functions that save me from writing everything 5 times for this particular program.