Create excelTranPerWriteTo.py
This allows for write to capability but I need to create functions to get rid of a lot of this code. Maybe each message box should be side a function with the same name. I think classes and functions and I could really make this a super simple program and one that could be scalable.